refactor(wave_topology): migrate to platform adapter fetchIssue
Closes #304
