+++ scripts/ci/migration-allowlist.txt
@@ -16,4 +16,3 @@ wave_finalize.ts
 wave_init.ts
 wave_previous_merged.ts
 wave_reconcile_mrs.ts
-wave_topology.ts
